Removed support for auto-casting integers to pointers. Closes #659

+++ src/main/java/dk/camelot64/kickc/model/operators/OperatorSizeOf.java
@@ -20,7 +20,7 @@ public class OperatorSizeOf extends OperatorUnary {
 
    @Override
    public SymbolType inferType(SymbolType operandType) {
-      return SymbolType.BYTE;
+      return SymbolType.WORD;
    }
 
 }
